refactor Books construtor to a Module, import Books from books_constructor to index.js
